#5 [FEAT] 단순 토큰 인증 기능 추가

+++ src/main/java/com/jaramgroupware/jgwauth/service/MemberAuthServiceImpl.java
@@ -5,11 +5,15 @@ import com.jaramgroupware.jgwauth.domain.redis.memberCache.MemberAuth;
 import com.jaramgroupware.jgwauth.domain.redis.memberCache.MemberAuthRepository;
 import com.jaramgroupware.jgwauth.dto.memberCache.servcieDto.MemberAuthAddRequestDto;
 import com.jaramgroupware.jgwauth.dto.memberCache.servcieDto.MemberAuthResponseDto;
+import com.jaramgroupware.jgwauth.dto.memberCache.servcieDto.TokenAuthAddRequestDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.Duration;
 import java.util.Optional;
 
 @Slf4j
@@ -18,12 +22,24 @@ import java.util.Optional;
 public class MemberAuthServiceImpl implements MemberAuthService {
 
     private final MemberAuthRepository memberAuthRepository;
+    private final RedisTemplate<String,String> redisTemplate;
+    private final ValueOperations<String,String> valueOperations;
 
     @Override
     @Transactional
     public boolean add(MemberAuthAddRequestDto memberCacheAddRequestDto) throws JsonProcessingException {
 
         memberAuthRepository.save(memberCacheAddRequestDto.toEntity());
+
+        return true;
+    }
+
+    @Override
+    @Transactional
+    public Boolean add(TokenAuthAddRequestDto tokenAuthAddRequestDto){
+
+        valueOperations.set("only_"+tokenAuthAddRequestDto.getToken(),tokenAuthAddRequestDto.getUid());
+        redisTemplate.expire("only_"+tokenAuthAddRequestDto.getToken(), Duration.ofMinutes(tokenAuthAddRequestDto.getTtl()));
 
         return true;
     }
@@ -44,14 +60,26 @@ public class MemberAuthServiceImpl implements MemberAuthService {
     }
 
     @Override
+    @Transactional(readOnly = true)
+    public Optional<String> findOnlyToken(String token) {
+
+        String uid = valueOperations.get("only_"+token);
+        if(uid == null) return Optional.empty();
+
+        return Optional.of(uid);
+    }
+
+    @Override
     @Transactional
     public boolean revoke(String token) {
         MemberAuth memberAuth = memberAuthRepository.findById(token)
                 .orElse(null);
 
-        if(memberAuth == null) return false;
+        String onlyToken = valueOperations.get("only_"+token);
 
-        memberAuthRepository.delete(memberAuth);
+        if(memberAuth != null) memberAuthRepository.delete(memberAuth);
+        if(onlyToken != null) redisTemplate.delete("only_"+token);
+
         return true;
     }
 }
